src/requirements.txt
matplotlib==3.9.2
numpy==2.1.1
Pillow==10.4.0
scikit_learn==1.5.2
